# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Performance Connect Event Intelligence Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List all discovery sources
      responses:
        "200":
          description: Discovery sources
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EventSource" }
    post:
      operationId: createSource
      tags: [sources]
      summary: Add a new discovery source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSourceInput"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventSource" }
  /sources/{id}:
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update a discovery source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSourceUpdate"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventSource" }
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Remove a discovery source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /sources/{id}/check:
    post:
      operationId: checkSource
      tags: [sources]
      summary: Scrape a discovery source now and queue any events found
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceCheckResult" }
        "404":
          description: Source not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EventSource:
      type: object
      required: [id, name, type, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [instagram, facebook, website, api, manual]
        url:
          type: ["string", "null"]
        active:
          type: boolean
        lastCheckedAt:
          type: ["string", "null"]
        eventsFound:
          type: integer
        createdAt:
          type: string
    SourceCheckResult:
      type: object
      required: [ok, message, eventsFound, duplicatesFound]
      properties:
        ok:
          type: boolean
        message:
          type: string
        eventsFound:
          type: integer
        duplicatesFound:
          type: integer
    EventSourceInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [instagram, facebook, website, api, manual]
        url:
          type: string
        active:
          type: boolean
          default: true
    EventSourceUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        active:
          type: boolean
...