# lib/api-zod/src/generated/types/sourceCheckResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Performance Connect Event Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface SourceCheckResult {
  ok: boolean;
  message: string;
  eventsFound: number;
  duplicatesFound: number;
}